update(auth): configured laravel sanctum for session cookie authentication
This improves security of the app by not storing the access_token in the local storage of the SPA
Users are not authenticated using the laravel_session HTTP-ONLY cookie

+++ backend/app/Http/Controllers/Api/V1/Auth/LogoutController.php
@@ -2,14 +2,24 @@
 
 namespace App\Http\Controllers\Api\V1\Auth;
 
-use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
+use Laravel\Sanctum\Http\Middleware\EnsureFrontendRequestsAreStateful;
 
 class LogoutController extends Controller
 {
     public function __invoke(Request $request)
     {
-        $request->user()->currentAccessToken()->delete();
+        if (EnsureFrontendRequestsAreStateful::fromFrontend($request)) {
+            Auth::guard('web')->logout();
+
+            $request->session()->invalidate();
+
+            $request->session()->regenerateToken();
+        } else {
+            $request->user()->currentAccessToken()->delete();
+        }
 
         return response()->noContent();
     }
